03_user_interfaces/02_single_container_az_func_vue_app/backend/requirements.txt
# Do not include azure-functions-worker in this file
# The Python Worker is managed by the Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions
pandas==2.2.2
pandocfilters==1.5.1
scikit-learn==1.5.2